renamed and reorganized request data helpers: services, models, store
