+++ src/app/dashboard/requests/[id]/confirm/ConfirmClient.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 import ConfirmUI from "./Confirm";
 import { useAuth } from "@/lib/auth";
-import { useRequestListStore } from "@/lib/stores/useRequestStore";
+import { useRequestListStore } from "@/lib/requests/store";
 import { toDateKey } from "@/lib/requestDates";
 
 interface ConfirmClientProps {
